+++ examples/llm/run.py
@@ -11,89 +11,17 @@ from morpheus.utils.logger import configure_logging
 from morpheus.pipeline import LinearPipeline
 from morpheus.config import Config
 
+from examples.llm.nemo import NemoLLMClient
+from examples.llm.mock import MockLLMClient
+from examples.llm.openai import OpenAICompletionLLMClient
+from examples.llm.llm_client import LLMClient
+
 import logging
 import cudf
 import mrc
 import typing
 import asyncio
-from abc import ABC, abstractmethod
-import nemollm
-import openai
 import os
-
-class LLMClient(ABC):
-
-    @abstractmethod
-    async def generate_async(self, prompt: str) -> str:
-        pass
-
-    @abstractmethod
-    async def generate_multiple_async(self, prompts: [str]) -> [str]:
-        pass
-
-class MockLLMClient(LLMClient):
-
-    _response: str
-
-    def __init__(self, response: str):
-        self._response = response
-
-    def _query(self, prompt: str) -> str:
-        return self._response
-
-    async def generate_async(self, prompt: str) -> str:
-        return self._query(prompt)
-
-    async def generate_multiple_async(self, prompts: [str]) -> [str]:
-        return [self._query(prompt) for prompt in prompts]
-
-class NemoLLMClient(LLMClient):
-
-    _connection: nemollm.NemoLLM
-    _model: str
-
-    def __init__(
-            self,
-            api_key: str | None = None,
-            org_id: str | None = None,
-            api_host: str | None = None,
-            model: str | None = None
-    ):
-        self._connection = nemollm.NemoLLM(api_key, org_id, api_host)
-        self._model = model or "gpt5b"
-
-    async def generate_async(self, prompt: str) -> str:
-        future = self._connection.generate(self._model, prompt, return_type="async")
-        result = await asyncio.wrap_future(future)
-        response = nemollm.NemoLLM.post_process_generate_response(result, return_text_completion_only=True)
-        return response
-
-    async def generate_multiple_async(self, prompts: [str]) -> [str]:
-      tasks = [asyncio.ensure_future(self.generate_async(prompt)) for prompt in prompts]
-      await asyncio.wait(tasks)
-      return [task.result() for task in tasks]
-
-class OpenAICompletionLLMClient(LLMClient):
-
-    _api_key: str
-    _model: str
-
-    def __init__(
-            self,
-            api_key: str | None = None,
-            model: str | None = None
-    ):
-        self._api_key = api_key
-        self._model = model or "ada"
-
-    async def generate_async(self, prompt: str) -> str:
-        result = await openai.Completion.acreate(model=self._model, prompt=prompt)
-        return result.choices[0].text
-
-    async def generate_multiple_async(self, prompts: [str]) -> [str]:
-      tasks = [asyncio.ensure_future(self.generate_async(prompt)) for prompt in prompts]
-      await asyncio.wait(tasks)
-      return [task.result() for task in tasks]
 
 class PromptCompletionStage(SinglePortStage):
